Sketch out module system Moves the prelude into a separate module. When a module is loaded, its toplevel items are placed in the runtime values table, somewhere after those from the prelude. The items from imported modules are also included. Items from these transitively-included modules are mangled with their containing module name to avoid name collisions.

+++ src/alloc.c
@@ -65,7 +65,7 @@ struct Allocator {
     struct ChunkId small[MAX_SMALL - 1];
     struct ChunkId hash[NUM_HASH];
     struct Chunk *chunks;
-    size_t nchunks;
+    uint32_t nchunks;
 
     paw_Alloc alloc;
     paw_Env *P;
